feat: get basic project functionality working

+++ src/main/java/student/TimeCard.java
@@ -7,7 +7,7 @@ public class TimeCard implements ITimeCard {
     private String employeeID;
     private double hoursWorked;
 
-    public TimeCard(String employeeID, float hoursWorked) {
+    public TimeCard(String employeeID, double hoursWorked) {
         this.employeeID = employeeID;
         this.hoursWorked = hoursWorked;
     }
